+++ PayPal/src/main/java/com/braintreepayments/api/paypal/PayPalRequest.kt
@@ -82,7 +82,7 @@ import org.json.JSONException
  * @property recurringBillingDetails Optional: Recurring billing product details.
  * @property recurringBillingPlanType Optional: Recurring billing plan type, or charge pattern.
  * @property userAction The call to action for the PayPal flow.
- * @property payPalCampaigns List of PayPal co-marketing campaigns to associate with the order.
+ * @property campaigns List of PayPal co-marketing campaigns to associate with the order.
  */
 abstract class PayPalRequest internal constructor(
     open val hasUserLocationConsent: Boolean,
@@ -106,7 +106,7 @@ abstract class PayPalRequest internal constructor(
     open var lineItems: List<PayPalLineItem> = emptyList(),
     open var userAction: PayPalPaymentUserAction = PayPalPaymentUserAction.USER_ACTION_DEFAULT,
     open var shouldOfferCredit: Boolean = false,
-    open var payPalCampaigns: List<PayPalCampaign> = emptyList()
+    open var campaigns: List<PayPalCampaign> = emptyList()
 ) : Parcelable {
 
     @Throws(JSONException::class)